+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # nevereverthrow
 
+## 0.2.1
+
+### Patch Changes
+
+- fix: Fix `ResultAsync.fromPromise` types for the new `finallyFn` ([#10](https://github.com/ieedan/nevereverthrow/pull/10))
+
+## 0.2.0
+
+### Minor Changes
+
+- feat: `ResultAsync.finally` ([#8](https://github.com/ieedan/nevereverthrow/pull/8))
+
 ## 0.1.0
 
 ### Minor Changes